Updated the srcdb package tests to use centralised testcontainers package - testcontainers interface and its implementations follows singleton pattern - implemented container registry to keep track of that and also added TerminateAllContainers() function using the registry - Using go:embed for storing the content of init schema files, and using the variable to make Start() load init file irrespective of wher the function is called from, in the project

+++ yb-voyager/test/containers/mysql_container.go
@@ -3,12 +3,14 @@ package testcontainers
 import (
 	"context"
 	"fmt"
+	"os"
 	"time"
 
 	"github.com/docker/go-connections/nat"
 	log "github.com/sirupsen/logrus"
 	"github.com/testcontainers/testcontainers-go"
 	"github.com/testcontainers/testcontainers-go/wait"
+	"github.com/yugabyte/yb-voyager/yb-voyager/src/utils"
 )
 
 type MysqlContainer struct {
@@ -17,7 +19,24 @@ type MysqlContainer struct {
 }
 
 func (ms *MysqlContainer) Start(ctx context.Context) (err error) {
+	if ms.container != nil {
+		utils.PrintAndLog("Mysql-%s container already running", ms.DBVersion)
+		return nil
+	}
+
+	// since these Start() can be called from anywhere so need a way to ensure that correct files(without needing abs path) are picked from project directories
+	tmpFile, err := os.CreateTemp(os.TempDir(), "mysql_schema.sql")
+	if err != nil {
+		return fmt.Errorf("failed to create temp schema file: %w", err)
+	}
+	defer tmpFile.Close()
+
+	if _, err := tmpFile.Write(mysqlInitSchemaFile); err != nil {
+		return fmt.Errorf("failed to write to temp schema file: %w", err)
+	}
+
 	req := testcontainers.ContainerRequest{
+		// TODO: verify the docker images being used are the correct/certified ones
 		Image:        fmt.Sprintf("mysql:%s", ms.DBVersion),
 		ExposedPorts: []string{"3306/tcp"},
 		Env: map[string]string{
@@ -29,7 +48,7 @@ func (ms *MysqlContainer) Start(ctx context.Context) (err error) {
 		WaitingFor: wait.ForListeningPort("3306/tcp").WithStartupTimeout(1 * time.Minute).WithPollInterval(1 * time.Second),
 		Files: []testcontainers.ContainerFile{
 			{
-				HostFilePath:      "./test_schemas/mysql_schema.sql",
+				HostFilePath:      tmpFile.Name(),
 				ContainerFilePath: "docker-entrypoint-initdb.d/mysql_schema.sql",
 				FileMode:          0755,
 			},
@@ -44,15 +63,22 @@ func (ms *MysqlContainer) Start(ctx context.Context) (err error) {
 }
 
 func (ms *MysqlContainer) Terminate(ctx context.Context) {
+	if ms == nil {
+		return
+	}
+
 	err := ms.container.Terminate(ctx)
 	if err != nil {
-		log.Errorf("faile to terminate mysql container: %v", err)
+		log.Errorf("failed to terminate mysql container: %v", err)
 	}
 }
 
 func (ms *MysqlContainer) GetHostPort() (string, int, error) {
-	ctx := context.Background()
+	if ms.container == nil {
+		return "", -1, fmt.Errorf("mysql container is not started: nil")
+	}
 
+	ctx := context.Background()
 	host, err := ms.container.Host(ctx)
 	if err != nil {
 		return "", -1, fmt.Errorf("failed to fetch host for mysql container: %w", err)
@@ -64,4 +90,8 @@ func (ms *MysqlContainer) GetHostPort() (string, int, error) {
 	}
 
 	return host, port.Int(), nil
+}
+
+func (ms *MysqlContainer) GetConfig() ContainerConfig {
+	return ms.ContainerConfig
 }
