+++ yb-voyager/test/containers/oracle_container.go
@@ -3,12 +3,14 @@ package testcontainers
 import (
 	"context"
 	"fmt"
+	"os"
 	"time"
 
 	"github.com/docker/go-connections/nat"
 	log "github.com/sirupsen/logrus"
 	"github.com/testcontainers/testcontainers-go"
 	"github.com/testcontainers/testcontainers-go/wait"
+	"github.com/yugabyte/yb-voyager/yb-voyager/src/utils"
 )
 
 type OracleContainer struct {
@@ -17,8 +19,25 @@ type OracleContainer struct {
 }
 
 func (ora *OracleContainer) Start(ctx context.Context) (err error) {
+	if ora.container != nil {
+		utils.PrintAndLog("Oracle-%s container already running", ora.DBVersion)
+		return nil
+	}
+
+	// since these Start() can be called from anywhere so need a way to ensure that correct files(without needing abs path) are picked from project directories
+	tmpFile, err := os.CreateTemp(os.TempDir(), "oracle_schema.sql")
+	if err != nil {
+		return fmt.Errorf("failed to create temp schema file: %w", err)
+	}
+	defer tmpFile.Close()
+
+	if _, err := tmpFile.Write(oracleInitSchemaFile); err != nil {
+		return fmt.Errorf("failed to write to temp schema file: %w", err)
+	}
+
 	// refer: https://hub.docker.com/r/gvenzl/oracle-xe
 	req := testcontainers.ContainerRequest{
+		// TODO: verify the docker images being used are the correct/certified ones (No license issue)
 		Image:        fmt.Sprintf("gvenzl/oracle-xe:%s", ora.DBVersion),
 		ExposedPorts: []string{"1521/tcp"},
 		Env: map[string]string{
@@ -30,7 +49,7 @@ func (ora *OracleContainer) Start(ctx context.Context) (err error) {
 		WaitingFor: wait.ForLog("DATABASE IS READY TO USE").WithStartupTimeout(2 * time.Minute).WithPollInterval(5 * time.Second),
 		Files: []testcontainers.ContainerFile{
 			{
-				HostFilePath:      "./test_schemas/oracle_schema.sql",
+				HostFilePath:      tmpFile.Name(),
 				ContainerFilePath: "docker-entrypoint-initdb.d/oracle_schema.sql",
 				FileMode:          0755,
 			},
@@ -45,15 +64,22 @@ func (ora *OracleContainer) Start(ctx context.Context) (err error) {
 }
 
 func (ora *OracleContainer) Terminate(ctx context.Context) {
+	if ora == nil {
+		return
+	}
+
 	err := ora.container.Terminate(ctx)
 	if err != nil {
-		log.Errorf("faile to terminate oracle container: %v", err)
+		log.Errorf("failed to terminate oracle container: %v", err)
 	}
 }
 
 func (ora *OracleContainer) GetHostPort() (string, int, error) {
-	ctx := context.Background()
+	if ora.container == nil {
+		return "", -1, fmt.Errorf("oracle container is not started: nil")
+	}
 
+	ctx := context.Background()
 	host, err := ora.container.Host(ctx)
 	if err != nil {
 		return "", -1, fmt.Errorf("failed to fetch host for oracle container: %w", err)
@@ -65,4 +91,8 @@ func (ora *OracleContainer) GetHostPort() (string, int, error) {
 	}
 
 	return host, port.Int(), nil
+}
+
+func (ora *OracleContainer) GetConfig() ContainerConfig {
+	return ora.ContainerConfig
 }
